Remove Dtype from object definition
